A rudimentary initial start
This will display a lot of trees. The fact that proper nouns are shown with NNP is a start and we can use similar tags for time. Since, the number of venues is limited, we can create a text file with all venues possible and check with that. Name of event can also be extracted though that will require some acrobatics. time and date is easily extracted.